+++ lib/components/screens/newgame.dart
@@ -1,6 +1,5 @@
 import 'package:EileMitWeile/components/box_victory.dart';
 import 'package:EileMitWeile/components/player.dart';
-import 'package:EileMitWeile/components/sprites/victory.dart';
 import 'package:flame/components.dart';
 import 'package:flame/experimental.dart';
 import 'package:flame/palette.dart';